start breadth first search on trees directory
